+++ cypress/cypress/support/index.d.ts
@@ -5,10 +5,10 @@ Permission to use, copy, modify, and distribute this software and its documentat
 The full terms of this copyright and license should always be found in the root directory of this software deliverable as "license.txt" and if these terms are not found with this software, please contact the USC Stevens Center for the full license.
 */
 
-declare namespace Cypress {
-  interface Chainable<Subject = any> {
-    matchImageSnapshot(value: string): Chainable<void>;
-    fill(value: string): Chainable<void>;
-  }
-  interface cy extends Chainable<undefined> {}
-}
+// declare namespace Cypress {
+//   interface Chainable<Subject = any> {
+//     matchImageSnapshot(value: string): Chainable<void>;
+//     fill(value: string): Chainable<void>;
+//   }
+//   interface cy extends Chainable<undefined> {}
+// }
